% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon-signed-rank-test.R
\name{wilcoxon_signed_rank_test}
\alias{wilcoxon_signed_rank_test}
\title{Tabulate wilcoxon signed rank significance test for multiple sets of differences}
\usage{
wilcoxon_signed_rank_test(dat, f, alternative = "two.sided", paired = TRUE,
  browse = FALSE)
}
\arguments{
\item{dat}{}

\item{f}{}

\item{alternative}{}

\item{paired}{}

\item{browse}{}

\item{var}{}
}
\description{
Tabulate wilcoxon signed rank significance test for multiple sets of differences
}
